Handle 1v1, 2v2, rotating and all regions

+++ app/src/main/java/com/nickoehler/brawlhalla/ranking/domain/StatDetail.kt
@@ -4,7 +4,7 @@ data class StatDetail(
 
     val brawlhallaId: Int,
     val name: String,
-    val xp: Double,
+    val xp: Int,
     val level: Int,
     val xpPercentage: Float,
     val games: Int,
